+++ web/src/js/__tests__/ducks/ui/filterSpec.tsx
@@ -2,6 +2,7 @@ import reducer, {
     FilterName,
     setFilter,
     setHighlight,
+    setHostname,
 } from "../../../ducks/ui/filter";
 
 jest.mock("../../../utils");
@@ -10,10 +11,18 @@ test("filter reducer", () => {
     expect(reducer(undefined, setFilter("foo"))).toEqual({
         [FilterName.Search]: "foo",
         [FilterName.Highlight]: "",
+        [FilterName.Hostname]: "",
     });
 
     expect(reducer(undefined, setHighlight("foo"))).toEqual({
         [FilterName.Search]: "",
         [FilterName.Highlight]: "foo",
+        [FilterName.Hostname]: "",
+    });
+
+    expect(reducer(undefined, setHostname("~d example.com"))).toEqual({
+        [FilterName.Search]: "",
+        [FilterName.Highlight]: "",
+        [FilterName.Hostname]: "~d example.com",
     });
 });