+++ web/src/js/components/Header.tsx
@@ -8,6 +8,7 @@ import { useAppDispatch, useAppSelector } from "../ducks";
 import FlowListMenu from "./Header/FlowListMenu";
 import OptionMenu from "./Header/OptionMenu";
 import FlowMenu from "./Header/FlowMenu";
+import RepeaterMenu from "./Header/RepeaterMenu";
 import type { Menu } from "./ProxyApp";
 import { Tab, setCurrent } from "../ducks/ui/tabs";
 
@@ -16,6 +17,7 @@ const tabs: { [key in Tab]: Menu } = {
     [Tab.FlowList]: FlowListMenu,
     [Tab.Options]: OptionMenu,
     [Tab.Flow]: FlowMenu,
+    [Tab.Repeater]: RepeaterMenu,
 };
 
 export default function Header() {
@@ -24,7 +26,12 @@ export default function Header() {
     const selectedFlows = useAppSelector((state) => state.flows.selected);
     const [wasFlowSelected, setWasFlowSelected] = useState(false);
 
-    const entries: Tab[] = [Tab.Capture, Tab.FlowList, Tab.Options];
+    const entries: Tab[] = [
+        Tab.Capture,
+        Tab.FlowList,
+        Tab.Repeater,
+        Tab.Options,
+    ];
     if (selectedFlows.length > 0) {
         entries.push(Tab.Flow);
     }