+++ web/src/js/ducks/ui/tabs.ts
@@ -6,6 +6,7 @@ export enum Tab {
     FlowList,
     Options,
     Flow,
+    Repeater,
 }
 
 const tabsSlice = createSlice({